test(web): move the #1385 aggregate tests into their own file
platforms/web/irrlicht.test.js was already 1407 lines before this
feature and CodeScene flags it for size and function count. The
aggregate's tests are self-contained - each builds the DOM it needs -
so they cost nothing by living in permissionsBanner.test.js, and
irrlicht.test.js goes back to byte-identical with main.
